refactor: Remove types module and update references
This commit:
- Removes the `mod types` declaration from `ring.rs`
- Deletes the now-empty `types.rs` file
- Fixes type inference and float conversion issues in various files
- Ensures type-safe error handling in topology.rs

+++ crates/core/src/tracing.rs
@@ -526,8 +526,8 @@ async fn send_to_metrics_server(
         }) => {
             let msg = PeerChange::added_connection_msg(
                 (&send_msg.tx != Transaction::NULL).then(|| send_msg.tx.to_string()),
-                (from_peer.clone().to_string(), from_loc.as_f64()),
-                (to_peer.clone().to_string(), to_loc.as_f64()),
+                (from_peer.clone().to_string(), from_loc.as_f64() as f64),
+                (to_peer.clone().to_string(), to_loc.as_f64() as f64),
             );
             ws_stream.send(Message::Binary(msg.into())).await
         }
